Main Page: Refactored login to seperate smaller chuncks and reworked the login flow. Was looping through access code refreshes without actually rechecking the access code each time.
Now tries the access code, if it fails, gets a new one, then resets the acesscode before trying again.

Timer:
Changed all of the ints to doubles to match the TimeSpan settings, updated some of the model to be TimeSpans instead of ints, and added Elapsed Time and Total Time counters.

+++ StreamWorks/StreamWorks/Models/Widgets/Timers/CreateStreamTimerModel.cs
@@ -2,17 +2,17 @@
 
 public class CreateStreamTimerModel
 {
-    public int StartingTime { get; set; } = 300;
+    public double StartingTime { get; set; } = 300;
     public bool ShowTotalCounts { get; set; } = false;
 
     // Twitch
-    public int TwitchFollowTime { get; set; } = 300;
-    public int TwitchTier1Time { get; set; } = 300;
-    public int TwitchTier2Time { get; set; } = 600;
-    public int TwitchTier3Time { get; set; } = 1000;
+    public double TwitchFollowTime { get; set; } = 300;
+    public double TwitchTier1Time { get; set; } = 300;
+    public double TwitchTier2Time { get; set; } = 600;
+    public double TwitchTier3Time { get; set; } = 1000;
 
     // Youtube
-    public int YouTubeLikeTime { get; set; } = 300;
-    public int YouTubeSubTime { get; set; } = 300;
+    public double YouTubeLikeTime { get; set; } = 300;
+    public double YouTubeSubTime { get; set; } = 300;
 
 }
